extract xcversion module to separate library file
